+++ NEWS.md
@@ -1,6 +1,23 @@
-# healthyR 0.1.1
+# healthyR (development version)
 
-# healthyR 0.1.0
+# healthyR 0.1.2
+Submit to CRAN
+* Add function
+1. ts_census_los_daily_tbl - Can get the census or length of stay given just two
+columns of data, admit/arrival date and discharge/departure date
+
+# healthyR 0.1.1
+* Add function(s)
+1. category_counts_tbl - Get counts of a specified column with a user specified
+grouping if provided
+2. named_item_list - Good for saving a tibble with many groups to different sheets
+of an excel file
+3. top_n_tbl - Good for getting a tibble of n records by some column selection
+4. ts_census_los_daily_tbl - Can compute the alos and census for data given simply
+the Admit Date and Discharge Date (discharge date can be null). Sometimes you want
+to see the average length of stay for those patients that are currently admitted
+
+## healthyR 0.1.0
 * Released to CRAN 12-03-2020
 
 ## healthyR 0.0.0.9006